// client-app/src/app/projects/+add-project/enums/project-types-name.enum.ts
export enum ProjectTypesSteps {
  AggregatedPpa = 'aggregated-ppas',
  BatteryStorage = 'battery-storage',
  CarbonOffset = 'carbon-offset-purchasing',
  CommunitySolar = 'community-solar',
  EacPurchasing = 'eac-purchasing',
  EfficiencyAudit = 'efficiency-audits-consulting',
  EfficiencyEquipmentMeasures = 'efficiency-equipment-measures',
  EmergingTechnologies = 'emerging-technologies',
  EvCharging = 'ev-charging-fleet-electrification',
  FuelCells = 'fuel-cells',
  OffsitePpa = 'offsite-power-purchase-agreement',
  OnsiteSolar = 'onsite-solar',
  RenewableRetail = 'renewable-retail-electricity',
  UtilityGreenTariff = 'utility-green-tariff'
}